# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://asi.admin.com/api/v1",
            "description": "Main Server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user with password",
                "description": "Validates user password and returns OAuth2 tokens",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthLoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "The user credentials were incorrect"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh access token",
                "description": "Refreshes access token by provided refresh token",
                "operationId": "authRefresh",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthRefreshRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "The refresh token is invalid"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user & revoke tokens",
                "description": "Revokes user access and refresh tokens",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AuthLoginRequest": {
                "required": [
                    "username",
                    "password"
                ],
                "properties": {
                    "username": {
                        "title": "username",
                        "description": "Username (email, min 6, max 60)",
                        "type": "string",
                        "example": "user@user.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "User Password (min 8, max 60)",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "AuthRefreshRequest": {
                "required": [
                    "refresh_token"
                ],
                "properties": {
                    "refresh_token": {
                        "title": "refresh_token",
                        "description": "Refresh Token (min 6, max 2048)",
                        "type": "string",
                        "example": "EXAMPLETOKEN"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use /auth/login endpoint to obtain a token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}